refactoring/2152 Use namespace-based APIs instead of Tenant-based APIs
Intermediate commit - quite a few tests failing.

* Merged NamespaceServiceIntegrationTest into NamespaceControllerIntegrationTest and deleted
* Updated reference in all tests
* Fixed an additional bug in NamespaceController
